+++ CHANGELOG.md
@@ -4,4 +4,5 @@
 
 * Adding a date selector.
 * Adding a preset selector.
+* Adding a frequentation chart to the location view.
 * Changed the `~` to `circa`.